Pin portable Phase 3 runtime dependencies

+++ crossfm/baselines.py
@@ -133,7 +133,7 @@ class QwenBinaryBaseline:
         if dtype is torch.bfloat16 and device.startswith("cuda") and not torch.cuda.is_bf16_supported():
             raise RuntimeError("bfloat16 was requested but is not supported by this CUDA device")
         self.model = AutoModelForCausalLM.from_pretrained(
-            model_id, revision=revision, torch_dtype=dtype, attn_implementation="sdpa"
+            model_id, revision=revision, dtype=dtype, attn_implementation="sdpa"
         ).to(device).eval()
         for parameter in self.model.parameters():
             parameter.requires_grad_(False)
